refactor(onboarding): components for QuestionBlock structural changes
