+++ frontend/src/pages/OnboardingPageV2/OnboardingPageV2.tsx
@@ -1,28 +1,28 @@
 import './OnboardingPageV2.styles.scss';
 
-import { Button, Col, Flex, Input, Layout, Row, Steps, Typography } from 'antd';
+import { Col, Layout, Row, Steps } from 'antd';
 import React, { useEffect, useRef, useState } from 'react';
 
+import QuestionBlock from './QuestionBlock';
+import { questions } from './questions';
+
 const { Header, Content } = Layout;
-const { Title, Text, Paragraph } = Typography;
-const { Search } = Input;
 
 interface OptionGroup {
-	id: string; // Unique identifier for the group
-	category: string; // The name of the category/group
-	items: string[]; // The options under the category
+	id: string;
+	category: string;
+	items: string[];
 }
 
-interface Question {
-	id: string; // Unique identifier for the question
-	heading: string; // The main question text
-	description: string; // A description providing context for the question
-	options: OptionGroup[]; // Array of grouped options
+export interface Question {
+	id: string;
+	title: string;
+	description: string;
+	options: OptionGroup[];
 	uiConfig?: {
-		showSearch?: boolean; // Whether to show the search input
-	}; // UI configuration options
+		showSearch?: boolean;
+	};
 }
-
 const setupStepItems = [
 	{
 		title: 'Org Setup',
@@ -40,156 +40,14 @@ const setupStepItems = [
 	},
 ];
 
-const questions: Question[] = [
-	{
-		id: 'question-block-1',
-		heading: 'Select your data source',
-		description:
-			'Select from a host of services to start sending data to SigNoz.',
-		options: [
-			{
-				id: 'category-1',
-				category: 'Application Monitoring',
-				items: ['Java', 'Python', 'Go', 'Node.js', 'Ruby', 'PHP'],
-			},
-			{
-				id: 'category-2',
-				category: 'Logs Monitoring',
-				items: [
-					'Kubernetes pods logs',
-					'Docker container logs',
-					'AWS CloudWatch',
-					'ElasticSearch',
-				],
-			},
-			{
-				id: 'category-3',
-				category: 'OpenTelemetry Collector',
-				items: ['OTLP HTTP', 'OTLP GRPC'],
-			},
-			{
-				id: 'category-4',
-				category: 'Other Metrics',
-				items: ['Host Metrics', 'Smart Agent Metrics'],
-			},
-			{
-				id: 'category-5',
-				category: 'AWS Metrics',
-				items: [
-					'AWS Application Logs',
-					'AWS EC2 Metrics',
-					'AWS ECS-Fargate',
-					'AWS Lambda',
-					'AWS Kinesis',
-				],
-			},
-			{
-				id: 'category-6',
-				category: 'Azure Metrics',
-				items: [
-					'Azure Function',
-					'Azure Container Logs',
-					'Azure App Service',
-					'AKS',
-					'Azure Database Metrics',
-					'Azure Blob Storage',
-				],
-			},
-			{
-				id: 'category-7',
-				category: 'Datadog',
-				items: ['Datadog Agent'],
-			},
-		],
-		uiConfig: {
-			showSearch: true,
-		},
-	},
-	{
-		id: 'question-block-2',
-		heading: 'Which Java Framework Do You Use?',
-		description: 'Descrption', // Description is inferred, might need adjustment
-		options: [
-			{
-				id: 'qb-2-category-1',
-				category: 'Java frameworks',
-				items: ['Sping Boot', 'Tomact', 'JBoss', 'Others'],
-			},
-		],
-	},
-	{
-		id: 'question-block-3',
-		heading: 'Select your data source',
-		description:
-			'Select from a host of services to start sending data to SigNoz.',
-		options: [
-			{
-				id: 'category-1',
-				category: 'Application Monitoring',
-				items: ['Java', 'Python', 'Go', 'Node.js', 'Ruby', 'PHP'],
-			},
-			{
-				id: 'category-2',
-				category: 'Logs Monitoring',
-				items: [
-					'Kubernetes pods logs',
-					'Docker container logs',
-					'AWS CloudWatch',
-					'ElasticSearch',
-				],
-			},
-			{
-				id: 'category-3',
-				category: 'OpenTelemetry Collector',
-				items: ['OTLP HTTP', 'OTLP GRPC'],
-			},
-			{
-				id: 'category-4',
-				category: 'Other Metrics',
-				items: ['Host Metrics', 'Smart Agent Metrics'],
-			},
-			{
-				id: 'category-5',
-				category: 'AWS Metrics',
-				items: [
-					'AWS Application Logs',
-					'AWS EC2 Metrics',
-					'AWS ECS-Fargate',
-					'AWS Lambda',
-					'AWS Kinesis',
-				],
-			},
-			{
-				id: 'category-6',
-				category: 'Azure Metrics',
-				items: [
-					'Azure Function',
-					'Azure Container Logs',
-					'Azure App Service',
-					'AKS',
-					'Azure Database Metrics',
-					'Azure Blob Storage',
-				],
-			},
-			{
-				id: 'category-7',
-				category: 'Datadog',
-				items: ['Datadog Agent'],
-			},
-		],
-		uiConfig: {
-			showSearch: false,
-		},
-	},
-];
-
 const questionnaireStepItems = questions.map((question) => ({
 	id: question.id,
-	title: question.heading,
+	title: question.title,
 	description: question.description,
 	options: question.options,
 	uiConfig: question.uiConfig,
 }));
+
 function OnboardingPageV2(): JSX.Element {
 	const [answers, setAnswers] = useState<string[]>(
 		new Array(questions.length).fill(''),
@@ -213,6 +71,9 @@ function OnboardingPageV2(): JSX.Element {
 		newAnswers[questionIndex] = option;
 		setAnswers(newAnswers);
 		setAnimatingOption(option);
+
+		console.log(`Question ${questionIndex + 1} Answer: ${option}`);
+		console.log('Current Form Output:', newAnswers);
 
 		setTimeout(() => setAnimatingOption(''), 600); // Reset animating state after animation completes
 
@@ -256,85 +117,18 @@ function OnboardingPageV2(): JSX.Element {
 						<Content>
 							<div className="setup-flow">
 								{questionnaireStepItems.map((item, index) => (
-									<div
+									<QuestionBlock
 										key={item.id}
-										className={`setup-flow__question-block ${
-											index <= currentQuestion ? 'setup-flow__question-block--active' : ''
-										}`}
-										ref={(el): void => {
-											questionRefs.current[index] = el;
-										}}
-									>
-										<Title level={4} className="setup-flow__content">
-											{item.title}
-										</Title>
-										<Paragraph className="setup-flow__description">
-											{item.description}
-										</Paragraph>
-										{item.uiConfig?.showSearch && (
-											<Search
-												placeholder="Search options"
-												onChange={handleSearch}
-												style={{ marginBottom: 16 }}
-											/>
-										)}
-										<div className="setup-flow__radio-buttons">
-											{item.options.map((group) => {
-												const filteredOptions = group.items.filter((option) =>
-													option.toLowerCase().includes(searchQuery),
-												);
-												if (filteredOptions.length === 0) return null;
-												return (
-													<React.Fragment key={group.id}>
-														{group.category && (
-															<div className="setup-flow__category">
-																{group.category} ({filteredOptions.length})
-															</div>
-														)}
-
-														{filteredOptions.map((option) => (
-															<label key={`${group.id}-option`} className="radio-label">
-																<input
-																	type="radio"
-																	name={`question-${index}`}
-																	value={option}
-																	checked={answers[index] === option}
-																	onChange={(): void => handleOptionChange(index, option)}
-																	className="setup-flow__radio-input"
-																/>
-																<span
-																	className={`setup-flow__radio-custom ${
-																		answers[index] === option
-																			? 'setup-flow__radio-custom--pulse setup-flow__radio-custom--selected'
-																			: ''
-																	} ${
-																		animatingOption === option
-																			? 'setup-flow__radio-custom--animating'
-																			: ''
-																	}`}
-																>
-																	<Text>{option}</Text>
-																</span>
-															</label>
-														))}
-													</React.Fragment>
-												);
-											})}
-											{item.uiConfig?.showSearch &&
-												item.options.every((group) =>
-													group.items.every(
-														(option) => !option.toLowerCase().includes(searchQuery),
-													),
-												) && (
-													<Flex gap={8} align="center" className="setup-flow__no-results">
-														<Text>No results found for &ldquo;{searchQuery}&rdquo;</Text>
-														<Button type="primary">
-															Tell our team, we will help you out
-														</Button>
-													</Flex>
-												)}
-										</div>
-									</div>
+										item={item}
+										index={index}
+										currentQuestion={currentQuestion}
+										answers={answers}
+										handleOptionChange={handleOptionChange}
+										animatingOption={animatingOption}
+										searchQuery={searchQuery}
+										handleSearch={handleSearch}
+										questionRefs={questionRefs}
+									/>
 								))}
 							</div>
 						</Content>
